Fix build errors after renaming.
