+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.2)
+cmake_minimum_required(VERSION 3.8)
 
 if(${CMAKE_VERSION} VERSION_LESS 3.11)
 	cmake_policy(VERSION ${CMAKE_MAJOR_VERSION}.${CMAKE_MINOR_VERSION})
@@ -14,6 +14,28 @@ project(hueplusplus LANGUAGES CXX)
 # options to set
 option(hueplusplus_TESTS "Build tests" OFF)
 
+option(CLANG_TIDY_FIX "Perform fixes for Clang-Tidy" OFF)
+find_program(CLANG_TIDY_EXE NAMES "clang-tidy" DOC "Path to clang-tidy executable")
+if(CLANG_TIDY_EXE)
+	if(CLANG_TIDY_FIX)
+		set(DO_CLANG_TIDY "${CLANG_TIDY_EXE}" "-fix")
+	else()
+		set(DO_CLANG_TIDY "${CLANG_TIDY_EXE}")
+	endif()
+endif()
+
+
+# Set default build type if none was specified
+set(default_build_type "Release")
+if(NOT CMAKE_BUILD_TYPE AND NOT CMAKE_CONFIGURATION_TYPES)
+	message(STATUS "Setting build type to '${default_build_type}' as none was specified")
+	set(CMAKE_BUILD_TYPE "${default_build_type}" CACHE STRING "Choose the type of build." FORCE)
+	# Set possible values for cmake-gui
+	set_property(CACHE CMAKE_BUILD_TYPE PROPERTY STRINGS
+		"Debug" "Release" "MinSizeRel" "RelWithDebInfo")
+endif()
+
+
 # get the correct installation directory for add_library() to work
 if(WIN32 AND NOT CYGWIN)
   set(DEF_INSTALL_CMAKE_DIR cmake)
@@ -25,7 +47,7 @@ set(INSTALL_CMAKE_DIR ${DEF_INSTALL_CMAKE_DIR} CACHE PATH "Installation director
 # target for uninstall
 if(NOT TARGET uninstall)
     configure_file(
-        "${CMAKE_CURRENT_SOURCE_DIR}/cmake_uninstall.cmake.in"
+        "${CMAKE_CURRENT_SOURCE_DIR}/cmake/cmake_uninstall.cmake.in"
         "${CMAKE_CURRENT_BINARY_DIR}/cmake_uninstall.cmake"
         IMMEDIATE @ONLY)
 
@@ -38,4 +60,9 @@ if (1 AND APPLE)
 	set(CMAKE_MACOSX_RPATH 1)
 endif()
 
-add_subdirectory(hueplusplus)
+add_subdirectory(src)
+
+# if the user decided to use tests add the subdirectory
+if(hueplusplus_TESTS)
+    add_subdirectory("test")
+endif()
