+++ include/hueplusplus/HueDeviceTypes.h
