+++ src/CMakeLists.txt
@@ -19,48 +19,43 @@ set(hueplusplus_SOURCES
 if(WIN32)
     set(hueplusplus_SOURCES
         ${hueplusplus_SOURCES}
-        ${CMAKE_CURRENT_SOURCE_DIR}/WinHttpHandler.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/WinHttpHandler.cpp PARENT_SCOPE
     )
 endif()
 # whereas on linux we want the LinHttpHandler
 if(UNIX)
     set(hueplusplus_SOURCES
         ${hueplusplus_SOURCES}
-        ${CMAKE_CURRENT_SOURCE_DIR}/LinHttpHandler.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/LinHttpHandler.cpp PARENT_SCOPE
     )
 endif()
 if(ESP_PLATFORM)
     set(hueplusplus_SOURCES
         ${hueplusplus_SOURCES}
-       ${CMAKE_CURRENT_SOURCE_DIR}/LinHttpHandler.cpp
+       ${CMAKE_CURRENT_SOURCE_DIR}/LinHttpHandler.cpp PARENT_SCOPE
     )
 endif()
 
+# Propagate to parent scope for tests
+set(hueplusplus_SOURCES ${hueplusplus_SOURCES} PARENT_SCOPE)
+
 # hueplusplus shared library
 add_library(hueplusplusshared SHARED ${hueplusplus_SOURCES})
-set_property(TARGET hueplusplusshared PROPERTY CXX_STANDARD 14)
-set_property(TARGET hueplusplusshared PROPERTY CXX_EXTENSIONS OFF)
-target_include_directories(hueplusplusshared PUBLIC $<BUILD_INTERFACE:"${CMAKE_CURRENT_SOURCE_DIR}/include"> $<INSTALL_INTERFACE:include/hueplusplus>)
+target_compile_features(hueplusplusshared PUBLIC cxx_std_14)
+target_include_directories(hueplusplusshared PUBLIC $<BUILD_INTERFACE:${hueplusplus_SOURCE_DIR}/include> $<INSTALL_INTERFACE:include>)
 install(TARGETS hueplusplusshared DESTINATION lib)
 
 # hueplusplus static library
 add_library(hueplusplusstatic STATIC ${hueplusplus_SOURCES})
-set_property(TARGET hueplusplusstatic PROPERTY CXX_STANDARD 14)
-set_property(TARGET hueplusplusstatic PROPERTY CXX_EXTENSIONS OFF)
+target_compile_features(hueplusplusstatic PUBLIC cxx_std_14)
 install(TARGETS hueplusplusstatic DESTINATION lib)
-target_include_directories(hueplusplusstatic PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/include")
+target_include_directories(hueplusplusstatic PUBLIC $<BUILD_INTERFACE:${hueplusplus_SOURCE_DIR}/include> $<INSTALL_INTERFACE:include>)
 install(FILES ${hueplusplus_HEADERS} DESTINATION include/hueplusplus)
 
 # Export the package for use from the build-tree
 # (this registers the build-tree with a global CMake-registry)
 export(PACKAGE hueplusplus)
 # Create the hueplusplus-config.cmake
-configure_file (hueplusplus-config.cmake.in "${PROJECT_BINARY_DIR}${CMAKE_FILES_DIRECTORY}/hueplusplus-config.cmake" @ONLY)
+configure_file ("${hueplusplus_SOURCE_DIR}/cmake/hueplusplus-config.cmake.in" "${hueplusplus_BINARY_DIR}${CMAKE_FILES_DIRECTORY}/hueplusplus-config.cmake" @ONLY)
 # Install hueplusplus-config.cmake
-install(FILES "${PROJECT_BINARY_DIR}${CMAKE_FILES_DIRECTORY}/hueplusplus-config.cmake" DESTINATION "${INSTALL_CMAKE_DIR}" COMPONENT dev)
-
-# if the user decided to use tests add the subdirectory
-if(hueplusplus_TESTS)
-    set(HuePlusPlus_INCLUDE_DIR ${CMAKE_CURRENT_SOURCE_DIR})
-    add_subdirectory("test")
-endif()
+install(FILES "${hueplusplus_BINARY_DIR}${CMAKE_FILES_DIRECTORY}/hueplusplus-config.cmake" DESTINATION "${INSTALL_CMAKE_DIR}" COMPONENT dev)
