+++ src/LinHttpHandler.cpp
@@ -20,7 +20,7 @@
     along with hueplusplus.  If not, see <http://www.gnu.org/licenses/>.
 **/
 
-#include "include/LinHttpHandler.h"
+#include "hueplusplus/LinHttpHandler.h"
 
 #include <chrono>
 #include <cstring>
