+++ test/test_Hue.cpp
@@ -30,8 +30,8 @@
 
 #include "testhelper.h"
 
-#include "../include/Hue.h"
-#include "../include/json/json.hpp"
+#include "hueplusplus/Hue.h"
+#include "json/json.hpp"
 #include "mocks/mock_HttpHandler.h"
 
 class HueFinderTest : public ::testing::Test
