+++ test/test_HueLight.cpp
@@ -25,9 +25,9 @@
 
 #include "testhelper.h"
 
-#include "../include/Hue.h"
-#include "../include/HueLight.h"
-#include "../include/json/json.hpp"
+#include "hueplusplus/Hue.h"
+#include "hueplusplus/HueLight.h"
+#include "json/json.hpp"
 #include "mocks/mock_HttpHandler.h"
 
 class HueLightTest : public ::testing::Test
