+++ test/test_SimpleColorTemperatureStrategy.cpp
@@ -1,5 +1,5 @@
 /**
-    \file test_ExtendedColorTemperatureStrategy.cpp
+    \file test_SimpleColorTemperatureStrategy.cpp
     Copyright Notice\n
     Copyright (C) 2017  Jan Rogall		- developer\n
     Copyright (C) 2017  Moritz Wirger	- developer\n
@@ -20,6 +20,7 @@
     along with hueplusplus.  If not, see <http://www.gnu.org/licenses/>.
 **/
 
+#include <iostream>
 #include <memory>
 #include <string>
 
@@ -28,12 +29,12 @@
 
 #include "testhelper.h"
 
-#include "../include/ExtendedColorTemperatureStrategy.h"
-#include "../include/json/json.hpp"
+#include "hueplusplus/SimpleColorTemperatureStrategy.h"
+#include "json/json.hpp"
 #include "mocks/mock_HttpHandler.h"
 #include "mocks/mock_HueLight.h"
 
-TEST(ExtendedColorTemperatureStrategy, setColorTemperature)
+TEST(SimpleColorTemperatureStrategy, setColorTemperature)
 {
     using namespace ::testing;
     std::shared_ptr<MockHttpHandler> handler(std::make_shared<MockHttpHandler>());
@@ -56,24 +57,23 @@ TEST(ExtendedColorTemperatureStrategy, setColorTemperature)
     prep_ret[2]["success"]["/lights/1/state/ct"] = 155;
     EXPECT_CALL(test_light, SendPutRequest(_, "/state", _)).Times(1).WillOnce(Return(prep_ret));
 
-    test_light.getState()["state"]["colormode"] = "ct";
     test_light.getState()["state"]["on"] = true;
     test_light.getState()["state"]["ct"] = 200;
-    EXPECT_EQ(true, ExtendedColorTemperatureStrategy().setColorTemperature(200, 4, test_light));
+    EXPECT_EQ(true, SimpleColorTemperatureStrategy().setColorTemperature(200, 4, test_light));
 
     test_light.getState()["state"]["on"] = false;
-    EXPECT_EQ(true, ExtendedColorTemperatureStrategy().setColorTemperature(155, 6, test_light));
+    EXPECT_EQ(true, SimpleColorTemperatureStrategy().setColorTemperature(155, 6, test_light));
 
     prep_ret[2]["success"]["/lights/1/state/ct"] = 153;
     EXPECT_CALL(test_light, SendPutRequest(_, "/state", _)).Times(1).WillOnce(Return(prep_ret));
-    EXPECT_EQ(true, ExtendedColorTemperatureStrategy().setColorTemperature(0, 6, test_light));
+    EXPECT_EQ(true, SimpleColorTemperatureStrategy().setColorTemperature(0, 6, test_light));
 
     prep_ret[2]["success"]["/lights/1/state/ct"] = 500;
     EXPECT_CALL(test_light, SendPutRequest(_, "/state", _)).Times(1).WillOnce(Return(prep_ret));
-    EXPECT_EQ(true, ExtendedColorTemperatureStrategy().setColorTemperature(600, 6, test_light));
+    EXPECT_EQ(true, SimpleColorTemperatureStrategy().setColorTemperature(600, 6, test_light));
 }
 
-TEST(ExtendedColorTemperatureStrategy, alertTemperature)
+TEST(SimpleColorTemperatureStrategy, alertTemperature)
 {
     using namespace ::testing;
     std::shared_ptr<MockHttpHandler> handler(std::make_shared<MockHttpHandler>());
@@ -86,47 +86,7 @@ TEST(ExtendedColorTemperatureStrategy, alertTemperature)
 
     test_light.getState()["state"]["colormode"] = "invalid";
     test_light.getState()["state"]["on"] = false;
-    EXPECT_EQ(false, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
-
-    EXPECT_CALL(test_light, setColorTemperature(_, _))
-        .Times(AtLeast(2))
-        .WillOnce(Return(false))
-        .WillRepeatedly(Return(true));
-    test_light.getState()["state"]["colormode"] = "hs";
-    test_light.getState()["state"]["on"] = true;
-    test_light.getState()["state"]["ct"] = 200;
-    test_light.getState()["state"]["sat"] = 100;
-    test_light.getState()["state"]["hue"] = 200;
-    EXPECT_EQ(false, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
-
-    EXPECT_CALL(test_light, alert()).Times(AtLeast(2)).WillOnce(Return(false)).WillRepeatedly(Return(true));
-    EXPECT_EQ(false, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
-
-    EXPECT_CALL(test_light, setColorHueSaturation(_, _, 1)).Times(AtLeast(2)).WillRepeatedly(Return(true));
-    EXPECT_EQ(true, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
-
-    EXPECT_CALL(test_light, OffNoRefresh(_)).Times(AtLeast(1)).WillRepeatedly(Return(true));
-    test_light.getState()["state"]["on"] = false;
-    EXPECT_EQ(true, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
-
-    EXPECT_CALL(test_light, setColorTemperature(_, _))
-        .Times(AtLeast(2))
-        .WillOnce(Return(false))
-        .WillRepeatedly(Return(true));
-    test_light.getState()["state"]["colormode"] = "xy";
-    test_light.getState()["state"]["on"] = true;
-    test_light.getState()["state"]["xy"][0] = 0.1;
-    test_light.getState()["state"]["xy"][1] = 0.1;
-    EXPECT_EQ(false, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
-
-    EXPECT_CALL(test_light, alert()).Times(AtLeast(2)).WillOnce(Return(false)).WillRepeatedly(Return(true));
-    EXPECT_EQ(false, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
-
-    EXPECT_CALL(test_light, setColorXY(_, _, 1)).Times(AtLeast(2)).WillRepeatedly(Return(true));
-    EXPECT_EQ(true, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
-
-    test_light.getState()["state"]["on"] = false;
-    EXPECT_EQ(true, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
+    EXPECT_EQ(false, SimpleColorTemperatureStrategy().alertTemperature(400, test_light));
 
     EXPECT_CALL(test_light, setColorTemperature(_, _))
         .Times(AtLeast(2))
@@ -134,15 +94,32 @@ TEST(ExtendedColorTemperatureStrategy, alertTemperature)
         .WillRepeatedly(Return(true));
     test_light.getState()["state"]["colormode"] = "ct";
     test_light.getState()["state"]["on"] = true;
-    test_light.getState()["state"]["on"] = true;
     test_light.getState()["state"]["ct"] = 200;
-    EXPECT_EQ(false, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
+    EXPECT_EQ(false, SimpleColorTemperatureStrategy().alertTemperature(400, test_light));
 
     EXPECT_CALL(test_light, alert()).Times(AtLeast(2)).WillOnce(Return(false)).WillRepeatedly(Return(true));
-    EXPECT_EQ(false, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
+    EXPECT_EQ(false, SimpleColorTemperatureStrategy().alertTemperature(400, test_light));
 
-    EXPECT_EQ(true, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
+    EXPECT_EQ(true, SimpleColorTemperatureStrategy().alertTemperature(400, test_light));
 
+    EXPECT_CALL(test_light, OffNoRefresh(_)).Times(AtLeast(1)).WillRepeatedly(Return(true));
     test_light.getState()["state"]["on"] = false;
-    EXPECT_EQ(true, ExtendedColorTemperatureStrategy().alertTemperature(400, test_light));
+    EXPECT_EQ(true, SimpleColorTemperatureStrategy().alertTemperature(400, test_light));
+}
+
+TEST(SimpleColorTemperatureStrategy, getColorTemperature)
+{
+    using namespace ::testing;
+    std::shared_ptr<MockHttpHandler> handler(std::make_shared<MockHttpHandler>());
+    EXPECT_CALL(
+        *handler, GETJson("/api/" + getBridgeUsername() + "/lights/1", nlohmann::json::object(), getBridgeIp(), 80))
+        .Times(AtLeast(1))
+        .WillRepeatedly(Return(nlohmann::json::object()));
+    MockHueLight test_light(handler);
+    EXPECT_CALL(test_light, refreshState()).Times(AtLeast(1)).WillRepeatedly(Return());
+
+    test_light.getState()["state"]["ct"] = 200;
+    EXPECT_EQ(200, SimpleColorTemperatureStrategy().getColorTemperature(test_light));
+    test_light.getState()["state"]["ct"] = 500;
+    EXPECT_EQ(500, SimpleColorTemperatureStrategy().getColorTemperature(static_cast<const HueLight>(test_light)));
 }
